// packages/components/src/ui/ToolbarTop/ToolbarTop.styles.d.ts
export declare const useStyles: () => {
  toolbarContainer: {
    backgroundColor: 'white';
    display: 'flex';
    flexDirection: 'row';
    alignItems: 'center';
    paddingLeft: import('@alveole/theme').Spacing;
    paddingRight: import('@alveole/theme').Spacing;
    paddingTop: import('@alveole/theme').Spacing;
    paddingBottom: import('@alveole/theme').Spacing;
  };
  compactLargetoolbarContainer: {
    paddingTop: import('@alveole/theme').Spacing;
    paddingBottom: import('@alveole/theme').Spacing;
    paddingLeft: import('@alveole/theme').Spacing;
    paddingRight: import('@alveole/theme').Spacing;
  };
  largeToolbarContainer: {
    paddingTop: number;
    paddingBottom: number;
    paddingLeft: number;
    paddingRight: number;
    flexDirection: 'column';
    alignItems: 'flex-start';
  };
  toolbarInformationWithBorder: {
    borderBottomWidth: number;
    borderLeftWidth: number;
    borderRightWidth: number;
    borderTopWidth: number;
    borderStyle: 'solid';
    borderColor: '#DEE3EC';
  };
  toolbarNavigation: {
    display: 'flex';
    justifyContent: 'center';
    gap: import('@alveole/theme').Spacing;
  };
  toolbarInformation: {
    display: 'flex';
    flexDirection: 'row';
    alignItems: 'center';
    gap: import('@alveole/theme').Spacing;
    flex: number;
  };
  toolbarInformationTitle: {
    display: 'flex';
    flexDirection: 'column';
    justifyContent: 'center';
    alignItems: 'flex-start';
    flex: number;
  };
  compactLargeInformations: {
    paddingBottom: number;
    justifyContent: 'center';
  };
  toolbarInformationTitleText: {
    color: '#151617';
    fontFamily:
      | 'Barlow-Light'
      | 'Barlow-Regular'
      | 'Barlow-Medium'
      | 'Barlow-SemiBold'
      | 'Barlow-Bold'
      | 'Inter-Light'
      | 'Inter-Regular'
      | 'Inter-Medium'
      | 'Inter-SemiBold'
      | 'Inter-Bold';
    fontSize: 14;
    lineHeight: 20;
    letterSpacing: 0;
  };
  largeInformationTitleText: {
    fontFamily:
      | 'Barlow-Light'
      | 'Barlow-Regular'
      | 'Barlow-Medium'
      | 'Barlow-SemiBold'
      | 'Barlow-Bold'
      | 'Inter-Light'
      | 'Inter-Regular'
      | 'Inter-Medium'
      | 'Inter-SemiBold'
      | 'Inter-Bold';
    fontSize: 32 | 40;
    lineHeight: 48 | 40;
    letterSpacing: 0;
  };
  toolbarInformationTitleSubText: {
    color: '#5F6571';
    fontFamily:
      | 'Barlow-Light'
      | 'Barlow-Regular'
      | 'Barlow-Medium'
      | 'Barlow-SemiBold'
      | 'Barlow-Bold'
      | 'Inter-Light'
      | 'Inter-Regular'
      | 'Inter-Medium'
      | 'Inter-SemiBold'
      | 'Inter-Bold';
    fontSize: 12;
    lineHeight: 20;
    letterSpacing: 0;
  };
  largeToolbarInformationTitleSubText: {
    fontFamily:
      | 'Barlow-Light'
      | 'Barlow-Regular'
      | 'Barlow-Medium'
      | 'Barlow-SemiBold'
      | 'Barlow-Bold'
      | 'Inter-Light'
      | 'Inter-Regular'
      | 'Inter-Medium'
      | 'Inter-SemiBold'
      | 'Inter-Bold';
    fontSize: 14;
    lineHeight: 20;
    letterSpacing: 0;
  };
  toolbarActions: {
    display: 'flex';
    flexDirection: 'row';
    justifyContent: 'flex-end';
    alignItems: 'center';
  };
  toolbarInformationContainer: {
    flex: number;
    alignItems: 'center';
    display: 'flex';
    flexDirection: 'row';
    justifyContent: 'space-between';
  };
};
//# sourceMappingURL=ToolbarTop.styles.d.ts.map